requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
sqlalchemy>=2.0.30
pydantic>=2.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator==2.3.0
bcrypt==4.0.1
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
